style(account): rename files for clarity
