+++ src-tauri/Cargo.toml
@@ -38,10 +38,8 @@ partial_derive = { path = "./libs/partial_derive" }
 uuid = { version = "1.11.0", features = ["v4", "v5", "fast-rng",  "macro-diagnostics"] }
 tauri-plugin-process = "2"
 tauri-plugin-opener = "2"
-json-web-token = "0.2.0"
 base64 = "0.22.1"
-jsonwebtoken = "9.3.1"
-image = "0.25.5"
+image = { version = "0.25.5", features = ["serde"] }
 rand = "0.9.0"
 
 [target."cfg(windows)".dependencies]
