+++ src-tauri/src/account/helpers/authlib_injector.rs
