+++ src-tauri/src/account/helpers/mod.rs
@@ -1,4 +1,4 @@
-pub mod auth_server;
-pub mod image;
+pub mod authlib_injector;
 pub mod offline;
+pub mod skin;
 pub mod storage;
