+++ src-tauri/src/utils/mod.rs
@@ -1,1 +1,2 @@
+pub mod image;
 pub mod sys_info;
